Refactor inference code
